requirements.txt
ftfy==6.3.1
h5py
huggingface_hub
matplotlib==3.5.1
ml_collections==0.1.1
numpy==1.21.6
opencv-python==4.5.5.64
openslide-python==1.2.0
pandas==1.3.5
regex
scikit-learn
scipy==1.7.3
tensorboardX==2.4
timm==1.0.15
torch-geometric==2.6.1
tqdm
transformers==4.49.0